hard-code bash to .do files, fixes #2
If you download and install just redo, it tries to find a capable shell and
creates a symlink to that in lib/redo. In abt, redo is never actually installed
and a proper shell is never detected. This is combination with the fact that
redo falls back to "sh" in your PATH, makes abt fail on systems where sh points
to dash for example (ubuntu does this).

Since bash is already used in ./abt, I'm just hard-coding it to the .do files as
well.